Added compile options, refactoring

+++ src/cli.cpp
@@ -66,7 +66,7 @@ void CLI::Start() {
 
 bool CLI::IsCodeCorrect(std::wstring &message) {
     bool code_is_correct = true;
-    for (int i = 0; i < message.length(); ++i) {
+    for (size_t i = 0; i < message.length(); ++i) {
         if ((int) (message[i] - '0') != 0 && (int) (message[i] - '0') != 1)
             code_is_correct = false;
     }
